backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.0.4
python-dotenv==1.0.0